(PC-6603) : Regroup recaptcha code into a single place
