+++ tests/routes/webapp/post_verify_id_check_licence_token_test.py
@@ -1,7 +1,7 @@
 from unittest.mock import MagicMock
 from unittest.mock import patch
 
-from pcapi.validation.routes.captcha import InvalidRecaptchaTokenException
+from pcapi.connectors.api_recaptcha import InvalidRecaptchaTokenException
 
 from tests.conftest import TestClient
